+++ src/utils/helpers/index.ts
@@ -7,3 +7,4 @@ export * from './grid-helpers';
 export * from './module-checks';
 export * from './naming-helpers';
 export * from './tag-helpers';
+export * from './tile-preview-helper';